ContentLibSubsystem.cpp: Fix unintended integer division
Both `e.Amount` and `i.Amount` are of `int32` type, but the result of
dividing these two numbers is stored in a `TArray` of `float` values.
Ensure float division is used by casting the first number to `float`.

Note that this function seems to be unused.

+++ Source/ContentLib/Private/ContentLibSubsystem.cpp
@@ -578,7 +578,7 @@ TArray<float> FFactoryGame_Recipe::GetIngredientsForProductRatio(const TSubclass
 		if(i.ItemClass != Item)
 			continue;
 		for (const auto e : Products) {
-			Array.Add(e.Amount / i.Amount);
+			Array.Add((float)e.Amount / i.Amount);
 		}
 	}
 	return Array;
